feat: extract shared hooks, sub-components, constants from TaskView/TaskConversationRenderer
- ROLE_COLORS → packages/web/src/lib/task-constants.ts
- parseGroupMessage + TaskMeta + ParsedGroupMessage → packages/web/src/lib/parse-group-message.ts
- Data-fetching and action logic → packages/web/src/hooks/useTaskViewData.ts
- HumanInputArea, TaskActionDialogs, TaskHeaderActions, TaskReviewBar → task-shared/
- V1 behavior unchanged; all existing 5088 tests pass
- Added 43 new unit tests for extracted modules
